commited

+++ POS/src/main/java/com/ust/pos/customer/service/impl/CustomerServiceImpl.java
@@ -1,15 +1,13 @@
 package com.ust.pos.customer.service.impl;
 
 import com.ust.pos.address.service.AddressService;
+import com.ust.pos.common.CommonService;
 import com.ust.pos.customer.service.CustomerService;
-import com.ust.pos.dto.AddressDto;
-import com.ust.pos.dto.CustomerDto;
-import com.ust.pos.dto.ShelfsDto;
+import com.ust.pos.dto.*;
 import com.ust.pos.model.Customer;
 import com.ust.pos.model.CustomerRepository;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -18,16 +16,21 @@ import java.lang.reflect.Type;
 import java.util.List;
 
 @Service
-public class CustomerServiceImpl implements CustomerService {
+public class CustomerServiceImpl extends CommonService implements CustomerService {
 
-    @Autowired
-    private CustomerRepository customerRepository;
+    private static final String CUSTOMER_WITH_IDENTIFIER = "Customer with identifier - ";
 
-    @Autowired
-    private ModelMapper modelMapper;
+    private final CustomerRepository customerRepository;
 
-    @Autowired
-    private AddressService addressService;
+    private final ModelMapper modelMapper;
+
+    private final AddressService addressService;
+
+    public CustomerServiceImpl(CustomerRepository customerRepository, ModelMapper modelMapper, AddressService addressService) {
+        this.customerRepository = customerRepository;
+        this.modelMapper = modelMapper;
+        this.addressService = addressService;
+    }
 
     @Override
     public CustomerDto findByIdentifier(String identifier) {
@@ -46,14 +49,22 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public CustomerDto save(CustomerDto customerDto) {
+        customerDto.setIdentifier(customerDto.getIdentifier().trim());
         String identifier = customerDto.getIdentifier();
         Customer existingCustomer = customerRepository.findByIdentifier(identifier);
         if (existingCustomer != null) {
-            customerDto.setMessage("Customer with identifier - " + identifier + " already exists");
+            if (!existingCustomer.isDeleted()) {
+                customerDto.setMessage(CUSTOMER_WITH_IDENTIFIER + identifier + " already exists");
+                customerDto.setSuccess(false);
+                return customerDto;
+            }
+            customerDto.setMessage(CUSTOMER_WITH_IDENTIFIER + identifier + " was previously deleted. " +
+                    "Please contact backend team to restore.");
             customerDto.setSuccess(false);
             return customerDto;
         }
         Customer customer = modelMapper.map(customerDto, Customer.class);
+        setAuditFields(customer,true);
         customerRepository.save(customer);
         AddressDto billingAddress = modelMapper.map(customerDto.getBillingAddress(), AddressDto.class);
         billingAddress.setIdentifier(customerDto.getIdentifier() + "_" + "Billing");
@@ -65,6 +76,8 @@ public class CustomerServiceImpl implements CustomerService {
         shippingAddress.setAddressType("Shipping");
         shippingAddress.setPhoneNo(customerDto.getIdentifier());
         addressService.save(shippingAddress);
+        customerDto.setSuccess(true);
+        customerDto.setMessage("Customer created successfully");
         return customerDto;
     }
 
@@ -73,11 +86,12 @@ public class CustomerServiceImpl implements CustomerService {
         String identifier = customerDto.getIdentifier();
         Customer existingCustomer = customerRepository.findByIdentifier(identifier);
         if (existingCustomer == null) {
-            customerDto.setMessage("Customer with identifier - " + identifier + " not found");
+            customerDto.setMessage(CUSTOMER_WITH_IDENTIFIER + identifier + " not found");
             customerDto.setSuccess(false);
             return customerDto;
         }
         modelMapper.map(customerDto, existingCustomer);
+        setAuditFields(existingCustomer,false);
         customerRepository.save(existingCustomer);
         List<AddressDto> addresses = addressService.findAllByPhoneNo(customerDto.getIdentifier());
         AddressDto billingAddress = customerDto.getBillingAddress();
@@ -91,31 +105,42 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public boolean delete(String identifier) {
-        customerRepository.deleteByIdentifier(identifier);
+        Customer customer = customerRepository.findByIdentifier(identifier);
+        if (customer == null) return false;
+        softDelete(customer);
+        setAuditFields(customer,false);
+        customerRepository.save(customer);
         addressService.delete(identifier);
         return true;
     }
 
     @Override
-    public List<CustomerDto> findAll(Pageable pageable) {
+    public WsDto<CustomerDto> findAll(Pageable pageable) {
         Type listType = new TypeToken<List<CustomerDto>>() {
         }.getType();
-        Page<Customer> customerPage = customerRepository.findAll(pageable);
-        return modelMapper.map(customerPage.getContent(), listType);
+        Page<Customer> customerPage = customerRepository.findByDeletedFalse(pageable);
+        WsDto<CustomerDto> customerDtoWsDto = new WsDto<>();
+        customerDtoWsDto.setDtoList(modelMapper.map(customerPage.getContent(), listType));
+        customerDtoWsDto.setTotalRecords(customerPage.getTotalElements());
+        customerDtoWsDto.setTotalPages(customerPage.getTotalPages());
+        customerDtoWsDto.setSizePerPage(pageable.getPageSize());
+        customerDtoWsDto.setPage(pageable.getPageNumber());
+        return customerDtoWsDto;
     }
 
     @Override
     public CustomerDto toggleStatus(String identifier) {
         Customer customer = customerRepository.findByIdentifier(identifier);
         customer.setStatus(!customer.isStatus());
+        setAuditFields(customer,false);
         customerRepository.save(customer);
         return modelMapper.map(customer, CustomerDto.class);
     }
 
     @Override
     public List<CustomerDto> findIfTrue() {
-        Type listType = new TypeToken<List<ShelfsDto>>() {
+        Type listType = new TypeToken<List<CustomerDto>>() {
         }.getType();
-        return modelMapper.map(customerRepository.findByStatusIsTrue(), listType);
+        return modelMapper.map(customerRepository.findByStatusIsTrueAndDeletedFalse(), listType);
     }
 }
